Conflicts:
	arch/arm/mach-exynos/Kconfig
	arch/arm/mach-exynos/board-universal5410-battery.c
	drivers/battery/max77803_charger.c
	drivers/cpufreq/exynos-ikcs-cpufreq.c
	drivers/hid/hid-ids.h
	drivers/media/video/exynos/fimg2d/fimg2d.h
	drivers/net/wireless/bcmdhd/Kconfig
	drivers/net/wireless/bcmdhd/Makefile
	drivers/net/wireless/bcmdhd/aiutils.c
	drivers/net/wireless/bcmdhd/bcmsdh_sdmmc.c
	drivers/net/wireless/bcmdhd/bcmsdh_sdmmc_linux.c
	drivers/net/wireless/bcmdhd/bcmutils.c
	drivers/net/wireless/bcmdhd/dhd.h
	drivers/net/wireless/bcmdhd/dhd_cdc.c
	drivers/net/wireless/bcmdhd/dhd_common.c
	drivers/net/wireless/bcmdhd/dhd_custom_gpio.c
	drivers/net/wireless/bcmdhd/dhd_custom_sec.c
	drivers/net/wireless/bcmdhd/dhd_linux.c
	drivers/net/wireless/bcmdhd/dhd_sdio.c
	drivers/net/wireless/bcmdhd/dhd_sec_feature.h
	drivers/net/wireless/bcmdhd/include/bcmdefs.h
	drivers/net/wireless/bcmdhd/include/bcmsdh_sdmmc.h
	drivers/net/wireless/bcmdhd/include/bcmutils.h
	drivers/net/wireless/bcmdhd/include/epivers.h
	drivers/net/wireless/bcmdhd/include/linux_osl.h
	drivers/net/wireless/bcmdhd/include/linuxver.h
	drivers/net/wireless/bcmdhd/linux_osl.c
	drivers/net/wireless/bcmdhd/wl_android.c
	drivers/net/wireless/bcmdhd/wl_cfg80211.c
	drivers/net/wireless/bcmdhd/wl_cfg80211.h
	drivers/net/wireless/bcmdhd/wl_cfgp2p.c

+++ drivers/sensorhub/factory/temphumidity_shtc1.c
@@ -69,8 +69,8 @@ err:
 
 static int convert_adc_to_temp(struct ssp_data *data, unsigned int adc)
 {
-	u8 low = 0, mid = 0;
-	u8 high;
+	int low = 0, mid = 0;
+	int high;
 
 	if (!data->cp_thm_adc_table || !data->cp_thm_adc_arr_size) {
 		/* using fake temp */
